+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -5,7 +5,7 @@
 import FlutterMacOS
 import Foundation
 
-import flutter_cef
+import flutter_cef_macos
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
   FlutterCefPlugin.register(with: registry.registrar(forPlugin: "FlutterCefPlugin"))